add film domain implementation
At this point we have a rough idea about the module that we want to build. By using BDD we will refine our knowledge about the domain and find out how it could work.
Since we want a fast feedback cycle, we will focus on domain without any IO. Integration tests would only slow us down (run in seconds instead of milliseconds), and we are not yet sure if our module design is correct, so no point in messing with database or controllers here.

Since our module needs to store data somewhere, we will create an in-memory repository. It's very, very simple, and require no special stubbing/mocking/libs.